requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-multipart
zxcvbn
exifread
python-nmap
scapy
stegano